+++ src/components/App.tsx
@@ -1,7 +1,9 @@
 import { env } from '@consts'
+import { FiveMins } from '@consts/time'
 import { addAllMintIds, getBalance, getMintsBalances, initDb } from '@db'
 import { fsInfo } from '@db/fs'
 import { l } from '@log'
+import type { INavigatorProps } from '@model/nav'
 import Navigator from '@nav/Navigator'
 import { NavigationContainer, NavigationContainerRef } from '@react-navigation/native'
 import { CustomErrorBoundary } from '@screens/ErrorScreen/ErrorBoundary'
@@ -9,14 +11,14 @@ import { ErrorDetails } from '@screens/ErrorScreen/ErrorDetails'
 import { FocusClaimProvider } from '@src/context/FocusClaim'
 import { KeyboardProvider } from '@src/context/Keyboard'
 import { NostrProvider } from '@src/context/Nostr'
-import { PinProvider } from '@src/context/Pin'
+import { PinCtx } from '@src/context/Pin'
 import { PrivacyProvider } from '@src/context/Privacy'
 import { PromptProvider } from '@src/context/Prompt'
 import { ThemeProvider, useThemeContext } from '@src/context/Theme'
-import { store } from '@store'
-import { STORE_KEYS } from '@store/consts'
+import { secureStore, store } from '@store'
+import { SECURESTORE_KEY, STORE_KEYS } from '@store/consts'
 import { dark, light } from '@styles'
-import { isErr } from '@util'
+import { isErr, isNull, isStr } from '@util'
 import { routingInstrumentation } from '@util/crashReporting'
 import { runRequestTokenLoop } from '@wallet'
 import * as SplashScreen from 'expo-splash-screen'
@@ -29,6 +31,15 @@ import * as Sentry from 'sentry-expo'
 import ClipboardModal from './ClipboardModal'
 import Toaster from './Toaster'
 // import { dropAllData } from '@src/storage/dev'
+
+interface ILockData {
+	mismatch: boolean
+	mismatchCount: number
+	locked: boolean
+	lockedCount: number
+	lockedTime: number,
+	timestamp: number
+}
 
 void SplashScreen.preventAutoHideAsync()
 
@@ -55,6 +66,22 @@ export default function App() {
 }
 
 function _App() {
+	// initial auth state
+	const [auth, setAuth] = useState<INavigatorProps>({
+		shouldSetup: false,
+		pinHash: ''
+	})
+	// app was longer than 5 mins in the background
+	const [bgAuth, setBgAuth] = useState(false)
+	// PIN mismatch state
+	const [attempts, setAttempts] = useState({
+		mismatch: false,
+		mismatchCount: 0,
+		locked: false,
+		lockedCount: 0,
+		lockedTime: 0,
+	})
+	const pinData = { attempts, setAttempts }
 	// i18next
 	const { t, i18n } = useTranslation(['common'])
 	// app ready to render content
@@ -98,14 +125,54 @@ function _App() {
 		}
 	}
 
+	// init pin auth data
+	const initAuth = async () => {
+		const [pinHash, shouldSetup] = await Promise.all([
+			secureStore.get(SECURESTORE_KEY),
+			store.get(STORE_KEYS.pinSkipped),
+		])
+		setAuth({
+			pinHash: isNull(pinHash) ? '' : pinHash,
+			shouldSetup: !isStr(shouldSetup) || !shouldSetup?.length
+		})
+		// check for pin attempts and app locked state
+		await handlePinForeground()
+	}
+
+	const handlePinForeground = async () => {
+		// check if app is locked
+		const now = Math.ceil(Date.now() / 1000)
+		const lockData = await store.getObj<ILockData>(STORE_KEYS.lock)
+		if (lockData) {
+			// set state acccording to lockData timestamp
+			const secsPassed = now - lockData.timestamp
+			const lockedTime = lockData.lockedTime - secsPassed
+			setAttempts({
+				...lockData,
+				mismatch: false,
+				lockedTime
+			})
+		}
+		// handle app was longer than 5 mins in the background
+		const bgTimestamp = await store.get(STORE_KEYS.bgCounter)
+		if (isStr(bgTimestamp) && bgTimestamp.length > 0) {
+			if (now - +bgTimestamp > FiveMins) {
+				setBgAuth(true)
+			}
+		}
+	}
+
 	// init
 	useEffect(() => {
-		void (async () => {
-			// init database
-			await initDB()
-			// init stored data
-			await initData()
-		})()
+		initDB().then(() => {
+			// ignore
+		}).catch(e => l(e))
+		initData().then(() => {
+			// ignore
+		}).catch(e => l(e))
+		initAuth().then(() => {
+			// ignore
+		}).catch(e => l(e))
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [])
 
@@ -115,14 +182,19 @@ function _App() {
 	// to ensure all initial requests are done before displaying content
 	return (
 		<ThemeProvider>
-			<PinProvider>
+			<PinCtx.Provider value={pinData}>
 				<FocusClaimProvider>
 					<PrivacyProvider>
 						<NostrProvider>
 							<NavContainer>
 								<PromptProvider>
 									<KeyboardProvider>
-										<Navigator />
+										<Navigator
+											shouldSetup={auth.shouldSetup}
+											pinHash={auth.pinHash}
+											bgAuth={bgAuth}
+											setBgAuth={setBgAuth}
+										/>
 										<StatusBar style="auto" />
 										<ClipboardModal />
 										<Toaster />
@@ -132,7 +204,7 @@ function _App() {
 						</NostrProvider>
 					</PrivacyProvider>
 				</FocusClaimProvider>
-			</PinProvider>
+			</PinCtx.Provider>
 		</ThemeProvider>
 	)
 }
